+++ android/app/src/main/kotlin/com/example/eilemitweile/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.eilemitweile
+package ch.atomikzopht.eilemitweile
 
 import io.flutter.embedding.android.FlutterActivity